+++ src/features/benefits/AbsorptionSection.tsx
