+++ src/features/benefits/BenefitsSection.tsx
