+++ src/features/benefits/index.ts
@@ -1,0 +1,2 @@
+export { default as BenefitsSection } from './BenefitsSection';
+export { default as AbsorptionSection } from './AbsorptionSection';
