+++ src/features/hero/components/ScrollIndicator.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { motion } from 'framer-motion';
+import { ChevronDown } from 'lucide-react';
+
+const ScrollIndicator: React.FC = () => {
+  return (
+    <motion.div
+      className="absolute bottom-8 left-1/2 transform -translate-x-1/2 text-center z-10"
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      transition={{ delay: 1, duration: 0.5 }}
+    >
+      <motion.div
+        animate={{ y: [0, 10, 0] }}
+        transition={{ duration: 1.5, repeat: Infinity }}
+        className="flex flex-col items-center text-juvelina-gold"
+      >
+        <ChevronDown size={30} />
+        <span className="text-sm text-gray-500 mt-1">Descubra Mais</span>
+      </motion.div>
+    </motion.div>
+  );
+};
+
+export default ScrollIndicator;
