+++ src/features/hero/components/ui/index.ts
@@ -1,0 +1,8 @@
+export { default as CreatorBadge } from './CreatorBadge';
+export { default as ScrollProgressBar } from './ScrollProgressBar';
+export { default as RecentActivityNotification } from './RecentActivityNotification';
+export { default as VisitorCounter } from './VisitorCounter';
+export { default as TikTokIcon } from './TikTokIcon';
+export { default as OptimizedImage } from './OptimizedImage';
+export { default as AnimatedButton } from './AnimatedButton';
+export { default as FireIcon } from './FireIcon';
