+++ src/features/pricing/index.ts
@@ -1,0 +1,2 @@
+export { default as PricingSection } from './PricingSection';
+export { default as ViralOfferSection } from './ViralOfferSection';
