+++ src/features/testimonials/index.ts
@@ -1,0 +1,4 @@
+export { default as TestimonialsSection } from './TestimonialsSection';
+export { default as UGCGallerySection } from './UGCGallerySection';
+export { default as VideoTestimonialsSection } from './VideoTestimonialsSection';
+export { default as ViralTestimonialsSection } from './ViralTestimonialsSection';
