+++ src/hooks/useScrollAnimation.tsx
